update UI
The most important changes are:
put search box and the result boxes into card views;
remove divider from the app;
adjust the relationship of the search box and result list

Now the 1st item in the list is dummy, used to occupy space used by the
search box. i.e. the real 1st result will not be covered by the search
box